Add examples of using ChannelConnection Publishing
Channel Connection Publishing is used by the third party providers
to emit events to Eventarc.
While this is similar to Custom Events is't using a different messages
and API endpoints.
So there is a benefit providing those samples to potential third party
providers.

I did not test all of those end to end, so addiitonal testing might be
needed.